+++ Utilities/CustomDateYearConvertor.cs
@@ -2,19 +2,23 @@
 using CsvHelper.Configuration;
 using CsvHelper;
 
-namespace Converters;
+namespace Utilities;
 
-public class CustomIntConverter : DefaultTypeConverter
+public class CustomDateYearConverter : DefaultTypeConverter
 {
     public override object ConvertFromString(string text, IReaderRow row, MemberMapData memberMapData)
     {
         if (text != "")
         {
-            return int.Parse(text);
+            var year = int.Parse(text);
+
+            var date = new DateOnly(year, 1, 1);
+            return date;
         }
         else
         {
-            return 0;
+            var date = new DateOnly(1, 1, 1);
+            return date;
         }
     }
 }
