+++ Utilities/CustomIntConverter.cs
@@ -2,23 +2,19 @@
 using CsvHelper.Configuration;
 using CsvHelper;
 
-namespace Converters;
+namespace Utilities;
 
-public class CustomDateYearConverter : DefaultTypeConverter
+public class CustomIntConverter : DefaultTypeConverter
 {
     public override object ConvertFromString(string text, IReaderRow row, MemberMapData memberMapData)
     {
         if (text != "")
         {
-            var year = int.Parse(text);
-
-            var date = new DateOnly(year, 1, 1);
-            return date;
+            return int.Parse(text);
         }
         else
         {
-            var date = new DateOnly(1, 1, 1);
-            return date;
+            return 0;
         }
     }
 }
